# optimize scripts controller tests