Speeds up the scripts controller tests by about 33%, mostly by moving object creation out of the setup method and into the test cases which need them. I did this because I was having to run these tests a lot while working on the new script update api.

## Before

```
Dave-MBP:~/src/cdo/dashboard (staging)$ time bundle exec spring testunit ./test/controllers/scripts_controller_test.rb
Running via Spring preloader in process 24873
Started with run options --seed 20219

/Users/dsb/src/cdo/dashboard/app/models/sections/section.rb:181: warning: constant ::TRUE is deprecated=---=---=---=---=--] 0% Time: 00:00:00,  ETA: ??:??:??
/Users/dsb/src/cdo/dashboard/app/models/sections/section.rb:181: warning: constant ::TRUE is deprecated                   ] 0% Time: 00:00:19,  ETA: 00:37:26
/Users/dsb/src/cdo/dashboard/app/models/sections/section.rb:181: warning: constant ::TRUE is deprecated                   ] 1% Time: 00:00:21,  ETA: 00:19:23
/Users/dsb/src/cdo/dashboard/app/models/sections/section.rb:181: warning: constant ::TRUE is deprecated                   ] 2% Time: 00:00:23,  ETA: 00:13:51
/Users/dsb/src/cdo/dashboard/app/models/sections/section.rb:181: warning: constant ::TRUE is deprecated                  ] 14% Time: 00:00:46,  ETA: 00:04:32
Resource object "all_resources" is missing course version and/or offering
Resource object "all_standards" is missing course version and/or offering
/Users/dsb/src/cdo/dashboard/app/models/sections/section.rb:181: warning: constant ::TRUE is deprecated                  ] 15% Time: 00:00:47,  ETA: 00:04:20
Resource object "all_code" is missing course version and/or offering
Resource object "all_resources" is missing course version and/or offering
Resource object "all_standards" is missing course version and/or offering
Resource object "all_vocabulary" is missing course version and/or offering
/Users/dsb/src/cdo/dashboard/app/models/sections/section.rb:181: warning: constant ::TRUE is deprecated                  ] 16% Time: 00:00:48,  ETA: 00:04:09
/Users/dsb/src/cdo/dashboard/app/models/sections/section.rb:181: warning: constant ::TRUE is deprecated                  ] 17% Time: 00:00:55,  ETA: 00:04:26
/Users/dsb/src/cdo/dashboard/app/models/sections/section.rb:181: warning: constant ::TRUE is deprecated                  ] 18% Time: 00:00:57,  ETA: 00:04:16
...
/Users/dsb/src/cdo/dashboard/app/models/sections/section.rb:181: warning: constant ::TRUE is deprecated=                 ] 85% Time: 00:03:27,  ETA: 00:00:36
/Users/dsb/src/cdo/dashboard/app/models/sections/section.rb:181: warning: constant ::TRUE is deprecated==                ] 86% Time: 00:03:29,  ETA: 00:00:33
/Users/dsb/src/cdo/dashboard/app/models/sections/section.rb:181: warning: constant ::TRUE is deprecated===               ] 87% Time: 00:03:30,  ETA: 00:00:31
Resource object "fake_name" is missing course version and/or offering============================================        ] 93% Time: 00:03:40,  ETA: 00:00:16
Resource object "fake_name" is missing course version and/or offering
Resource object "fake_name" is missing course version and/or offering
Resource object "fake_name_1" is missing course version and/or offering
Resource object "fake_name_1" is missing course version and/or offering
Resource object "fake_name_1" is missing course version and/or offering
/Users/dsb/src/cdo/dashboard/app/models/sections/section.rb:181: warning: constant ::TRUE is deprecated===========       ] 94% Time: 00:03:42,  ETA: 00:00:14
Resource object "fake_name" is missing course version and/or offering
Resource object "fake_name" is missing course version and/or offering
Resource object "fake_name" is missing course version and/or offering
Resource object "fake_name_1" is missing course version and/or offering
Resource object "fake_name_1" is missing course version and/or offering
Resource object "fake_name_1" is missing course version and/or offering
Resource object "fake_name_2" is missing course version and/or offering
Resource object "fake_name_2" is missing course version and/or offering
Resource object "fake_name_2" is missing course version and/or offering
/Users/dsb/src/cdo/dashboard/app/models/sections/section.rb:181: warning: constant ::TRUE is deprecated================  ] 99% Time: 00:03:50,  ETA: 00:00:03
  103/103: [============================================================================================================] 100% Time: 00:03:52, Time: 00:03:52

Finished in 232.07734s
103 tests, 255 assertions, 0 failures, 0 errors, 0 skips
bundle exec spring testunit ./test/controllers/scripts_controller_test.rb  0.34s user 0.15s system 0% cpu 3:57.41 total
```

## after

```
Dave-MBP:~/src/cdo/dashboard (optimize-scripts-controller-tests)$ time bundle exec spring testunit ./test/controllers/scripts_controller_test.rb
Running via Spring preloader in process 23947
Started with run options --seed 44890

Resource object "all_resources" is missing course version and/or offering                                                ] 24% Time: 00:00:36,  ETA: 00:01:53
Resource object "all_standards" is missing course version and/or offering
Resource object "all_code" is missing course version and/or offering                                                     ] 25% Time: 00:00:36,  ETA: 00:01:49
Resource object "all_resources" is missing course version and/or offering
Resource object "all_standards" is missing course version and/or offering
Resource object "all_vocabulary" is missing course version and/or offering
Resource object "fake_name" is missing course version and/or offering============================================        ] 93% Time: 00:02:24,  ETA: 00:00:11
Resource object "fake_name" is missing course version and/or offering
Resource object "fake_name" is missing course version and/or offering
Resource object "fake_name_1" is missing course version and/or offering
Resource object "fake_name_1" is missing course version and/or offering
Resource object "fake_name_1" is missing course version and/or offering
Resource object "fake_name" is missing course version and/or offering=============================================       ] 94% Time: 00:02:25,  ETA: 00:00:09
Resource object "fake_name" is missing course version and/or offering
Resource object "fake_name" is missing course version and/or offering
Resource object "fake_name_1" is missing course version and/or offering
Resource object "fake_name_1" is missing course version and/or offering
Resource object "fake_name_1" is missing course version and/or offering
Resource object "fake_name_2" is missing course version and/or offering
Resource object "fake_name_2" is missing course version and/or offering
Resource object "fake_name_2" is missing course version and/or offering
  103/103: [============================================================================================================] 100% Time: 00:02:31, Time: 00:02:31

Finished in 151.07421s
103 tests, 255 assertions, 0 failures, 0 errors, 0 skips
bundle exec spring testunit ./test/controllers/scripts_controller_test.rb  0.34s user 0.16s system 0% cpu 2:34.50 total
```